refactor vertex controls into separate components & iterate
