+++ src/components/scene/vertex_controls/index.ts
@@ -1,0 +1,3 @@
+export { default as GradientVertexColorControls } from './GradientVertexColorControls';
+export { default as HslVertexColorControls } from './HslVertexColorControls';
+export { default as PickVertexColorControls } from './PickVertexColorControls';
